Introduce Halton sampler also for progressive photon mapping.

+++ sources/halton.cc
@@ -59,7 +59,7 @@ namespace {
         7793, 7817, 7823, 7829, 7841, 7853, 7867, 7873, 7877, 7879, 7883, 7901, 7907, 7919
     };
 
-    void shuffle(int* p, int d, Random& rand) {
+    void shuffle(long long* p, int d, Random& rand) {
         for (int i = 0; i < d; i++) {
             const int r = rand.nextInt(d - i);
             std::swap(p[i], p[i + r]);
@@ -69,7 +69,7 @@ namespace {
 
 Halton::Halton(int dim, bool isPermute, unsigned int seed) 
     : dims(dim)
-    , usedSamples(0)
+    , usedSamples(0ll)
     , bases()
     , permute()
 {
@@ -84,7 +84,7 @@ Halton::Halton(int dim, bool isPermute, unsigned int seed)
     }
 
     permute.resize(sumBases);
-    int* p = &permute[0];
+    long long* p = &permute[0];
     for (int i = 0; i < dims; i++) {
         for (int k = 0; k < bases[i]; k++) {
             p[k] = k;
@@ -122,7 +122,7 @@ void Halton::request(RandomSequence& rseq, int n) {
     Assertion(n <= dims, "Requested samples are too many!!");
     rseq.resize(n);
 
-    int* p = &permute[0];
+    long long* p = &permute[0];
     for (int i = 0; i < n; i++) {
         rseq.set(i, radicalInverse(usedSamples, bases[i], p));
         p += bases[i];
@@ -131,12 +131,12 @@ void Halton::request(RandomSequence& rseq, int n) {
     rseq.reset();
 }
 
-double Halton::radicalInverse(int n, int base, const int* p) const {
+double Halton::radicalInverse(long long n, long long base, const long long* p) const {
     double val = 0.0;
     double invBase = 1.0 / base;
     double invBi = invBase;
     while (n > 0) {
-        int d_i = p[n % base];
+        long long d_i = p[n % base];
         val += d_i * invBi;
         invBi *= invBase;
         n /= base;
